TP_Microservices_Mini_Projet.md/mini-projet/person-service/requirements.txt
flask~=2.1.2
Werkzeug==2.2.3
Flask-SQLAlchemy~=2.5.1
SQLAlchemy<2.0
Flask-Cors~=3.0.10
PyJWT~=2.4.0
gunicorn~=20.1.0
